QuadraticParameters
