+++ rust/ommx/src/random.rs
@@ -7,7 +7,10 @@ use proptest::{
 };
 
 mod linear;
+mod quadratic;
+
 pub use linear::*;
+pub use quadratic::*;
 
 /// Get random object based on [`Arbitrary`] trait with its [`Arbitrary::Parameters`].
 pub fn random<T: Arbitrary>(rng: proptest::test_runner::TestRng, parameters: T::Parameters) -> T {
@@ -30,6 +33,7 @@ pub fn random_deterministic<T: Arbitrary>(parameters: T::Parameters) -> T {
     tree.current()
 }
 
+/// Strategy for generating arbitrary coefficients.
 pub fn arbitrary_coefficient() -> BoxedStrategy<f64> {
     prop_oneof![Just(0.0), Just(1.0), Just(-1.0), -1.0..1.0].boxed()
 }
